Update gps_information.md
Hi there! I spent some time reverse engineering how to sync the GPS location with the camera. I tried using the documentation you provided, but it is partly incorrect. The correct format seems to be omitting the timezone info (all timestamps are in UTC), and for whatever reason the year is in little endian order.

I cross-checked with the values returned from the camera after syncing using the Ricoh's app, and that lines up. Yes, I know how weird that sounds 🙃